fix: change data size

+++ test/test_comm.hpp
@@ -1,17 +1,16 @@
 /**
  * @file test_comm.hpp
  * @author Teruaki Ata (PFA03027@nifty.com)
- * @brief 
+ * @brief
  * @version 0.1
  * @date 2025-07-20
- * 
+ *
  * @copyright Copyright (c) 2025, Teruaki Ata (PFA03027@nifty.com)
- * 
+ *
  */
 
 #ifndef TEST_COMM_HPP_
 #define TEST_COMM_HPP_
-
 
 class NonTrivialType {
 public:
@@ -19,15 +18,15 @@ public:
 	{
 		delete p_value;   // Clean up dynamically allocated memory
 	}
-	NonTrivialType( int value = 42 )
-	  : p_value( new int( value ) ) {}
+	NonTrivialType( size_t value = 42 )
+	  : p_value( new size_t( value ) ) {}
 
 	NonTrivialType( const NonTrivialType& other )
-	  : p_value( new int( *other.p_value ) )   // Deep copy
+	  : p_value( new size_t( *other.p_value ) )   // Deep copy
 	{
 	}
 	NonTrivialType( NonTrivialType&& other ) noexcept
-	  : p_value( new int( *other.p_value ) )   // Deep copy
+	  : p_value( new size_t( *other.p_value ) )   // Deep copy
 	{
 	}
 
@@ -36,8 +35,8 @@ public:
 		if ( this == &other ) {
 			return *this;   // Handle self-assignment
 		}
-		delete p_value;                        // Clean up existing resource
-		p_value = new int( *other.p_value );   // Deep copy
+		delete p_value;                           // Clean up existing resource
+		p_value = new size_t( *other.p_value );   // Deep copy
 		return *this;
 	}
 	NonTrivialType& operator=( NonTrivialType&& other ) noexcept
@@ -45,18 +44,18 @@ public:
 		if ( this == &other ) {
 			return *this;   // Handle self-assignment
 		}
-		delete p_value;                        // Clean up existing resource
-		p_value = new int( *other.p_value );   // Deep copy
+		delete p_value;                           // Clean up existing resource
+		p_value = new size_t( *other.p_value );   // Deep copy
 		return *this;
 	}
 
-	int get_value() const
+	size_t get_value() const
 	{
 		return p_value ? *p_value : 0;   // Return value or 0 if p_value is null
 	}
 
 private:
-	int* p_value;   //!< Pointer to an integer value, which is non-trivial due to dynamic memory allocation.
+	size_t* p_value;   //!< Pointer to an integer value, which is non-trivial due to dynamic memory allocation.
 };
 
 #endif   // TEST_COMM_HPP_
